Code cleanup, rename to match mental model
